data-consumer/requirements.txt
kafka-python
pandas
sqlalchemy
psycopg2-binary

requests
